Added ros bag feature in launch file
In this commit I have adde ros2 bag feature to record all the topics and save it in
the result folder. I have also modified Luanch file to pass argument to record it
or not for 15 sec.
Updated Readme with instructions to record and play the bag file and steps to verify it.